+++ app/src/main/java/eu/faircode/email/ActivityError.java
@@ -44,6 +44,7 @@ public class ActivityError extends ActivityBase {
     private Button btnPassword;
     private ImageButton ibSetting;
     private ImageButton ibInfo;
+    private Button btnReload;
     private Button btnSupport;
 
     @Override
@@ -60,6 +61,7 @@ public class ActivityError extends ActivityBase {
         btnPassword = findViewById(R.id.btnPassword);
         ibSetting = findViewById(R.id.ibSetting);
         ibInfo = findViewById(R.id.ibInfo);
+        btnReload = findViewById(R.id.btnReload);
         btnSupport = findViewById(R.id.btnSupport);
 
         load();
@@ -100,7 +102,9 @@ public class ActivityError extends ActivityBase {
         tvMessage.setMovementMethod(LinkMovementMethodCompat.getInstance());
         tvMessage.setText(message);
 
-        tvCertificate.setVisibility(isCertificateException ? View.VISIBLE : View.GONE);
+        tvCertificate.setVisibility(
+                isCertificateException && !SSLHelper.customTrustManager()
+                        ? View.VISIBLE : View.GONE);
 
         boolean password = (auth_type == ServiceAuthenticator.AUTH_TYPE_PASSWORD);
 
@@ -174,6 +178,15 @@ public class ActivityError extends ActivityBase {
             @Override
             public void onClick(View view) {
                 Helper.viewFAQ(view.getContext(), isCertificateException ? 4 : faq);
+            }
+        });
+
+        btnReload.setVisibility(account > 0 ? View.VISIBLE : View.GONE);
+        btnReload.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                ServiceSynchronize.reload(v.getContext(), account, true, "retry");
+                finish();
             }
         });